pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pytelegrambotapi>=4.26.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40",
    "telebot>=0.0.5",
    "werkzeug>=3.1.3",
    "alembic>=1.15.2",
    "trafilatura>=2.0.0",
    "python-dotenv>=1.1.0",
]
